Create HeadlessMixedObjectScanner
HeadlessMixedObjectScanner assumes that the object has no header or
class pointer.  MixedObjectScanner is now built on
HeadlessMixedObjectScanner.
